+++ src/extension.ts
@@ -2,61 +2,22 @@ import * as vscode from 'vscode'
 import { buildCommands } from './commands'
 import { consumeSelectedBlock, convertToDefinitions } from './common'
 import { buildEvents } from './events'
-import { createNewFish, getNewFishPlaceHolder, processFishName } from './newFish'
+import { createExportForAllFishes, createFishExport } from './exportFish'
+import { createNewFish } from './newFish'
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext): void {
-  const newFish = vscode.commands.registerCommand('actyx.newFish', async () => {
-    const { window } = vscode
-    const editor = window.activeTextEditor
-    if (editor) {
-      const placeHolder = getNewFishPlaceHolder(editor)
-      window.showInputBox({ placeHolder }).then(fishName => {
-        // validate user input as fish name
-        const processedFishName = processFishName(fishName, placeHolder)
-        if (!processedFishName) {
-          vscode.window.showInformationMessage('The fish name is mandatory')
-          return
-        }
+  const newFish = vscode.commands.registerCommand('actyx.newFish', async () => createNewFish())
 
-        // if there are more than two lines, get approvement from user to delete content
-        if (editor.document.lineCount > 2) {
-          window
-            .showWarningMessage<vscode.MessageItem>(
-              'Your current file is not empty. This command will overwrite your content',
-              { modal: true },
-              { title: 'Overwrite', isCloseAffordance: false },
-              { title: 'Cancel', isCloseAffordance: true },
-            )
-            .then(result => {
-              if (result && result.title === 'Overwrite') {
-                createNewFish(editor, processedFishName).then(_ => {
-                  window.showInformationMessage(
-                    'Fish created. Continue with "Actyx: create event" and "Actyx: create commands" ',
-                  )
-                  // save when the file is not saved
-                  if (editor.document.isUntitled) {
-                    editor.document.save()
-                  }
-                }).catch(console.error)
-              } else {
-                vscode.window.showInformationMessage('Create new fish is canceled by user')
-              }
-            })
-        } else {
-          // tslint:disable-next-line: no-floating-promises
-          createNewFish(editor, processedFishName).then(_ => {
-            window.showInformationMessage(
-              'Fish created. Continue with "Actyx: create event" and "Actyx: create commands" ',
-            )
-            // save when the file is not saved
-            if (editor.document.isUntitled) {
-              editor.document.save()
-            }
-          })
-        }
-      })
+  const exportFish = vscode.commands.registerCommand('actyx.exportFish', async () => {
+    const editor = vscode.window.activeTextEditor
+    if (editor) {
+      if (editor.document.fileName.includes('index.ts')) {
+        await createExportForAllFishes(editor)
+      } else {
+        await createFishExport(editor)
+      }
     }
   })
 
@@ -85,6 +46,7 @@ export function activate(context: vscode.ExtensionContext): void {
   })
 
   context.subscriptions.push(newFish)
+  context.subscriptions.push(exportFish)
   context.subscriptions.push(eventsCommand)
   context.subscriptions.push(commandsCommand)
 }