+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Change Log
 
-All notable changes to the "actyx-pond" extension will be documented in this file.
+## version 0.2.0
 
-## [released]
+* `export fish`
+  * export single fish from definition file
+  * export all fishes next to the index.ts file
+* improvements
+  * LF and CRLF compatible
+  * lazy/incomplete selection is now supported
+  * auto save after new Fish
+* add keybindings
+  * `Alt+a n` new fish in current file
+  * `Alt+a x` export fish definition
+  * `Alt+a e` create events for selected definition
+  * `Alt+a c` create Commands for selected definition
 
-version 0.1.0
+## version 0.1.0
 
 * `new fish`
   * create new fish